Set rules, how HTML and CSS code should written like
Format the code in index.html like so:
<div>

  <div>
    <div />
    <div />
  </div>

  <div>

    <div>
      <div>
        <div />
      </div>
    </div>

    <div>

      <div />
      <div />

      <div>
        <div />
      </div>

    </div>

  </div>

</div>

Reorder the elements in <head> so, that at te beginning is <meta>, then <title> and <link>. The link to style.css is at the end.

If the href attribute of a link is empty, give it the value # (hashtag).

Remove justify-content property from '.row' and '.wrapper'. Create rules for columns with flex from 1 to 11.

Give the colums at index.html the right classes and remove the class profil_img.

[O
At style.css, order the rules similary like at index.html. Give a space after the comment /* Contact */. Change selector .first_section to header at Media.